+++ src/test/java/chess/domain/piece/KnightTest.java
@@ -1,11 +1,10 @@
-package chess.domain;
+package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
 import java.util.List;
-import chess.domain.piece.Knight;
-import chess.domain.piece.Team;
+import chess.domain.Coordinate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
