+++ src/test/java/chess/domain/piece/RookTest.java
@@ -1,29 +1,27 @@
-package chess.domain;
+package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
 import java.util.List;
-import chess.domain.piece.Queen;
-import chess.domain.piece.Rook;
-import chess.domain.piece.Team;
+import chess.domain.Coordinate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class QueenTest {
+class RookTest {
 
     @DisplayName("생성 테스트")
     @Test
     void create() {
-        assertThatCode(() -> new Queen(Team.WHITE))
+        assertThatCode(() -> new Rook(Team.WHITE))
                 .doesNotThrowAnyException();
     }
 
-    @DisplayName("퀸은 대각과 가로, 세로로 제한없이 움직일 수 있다.")
+    @DisplayName("룩은 가로, 세로로 제한없이 움직일 수 있다.")
     @Test
     void findAllPossibleCoordinate() {
         Coordinate start = new Coordinate(4, 'd');
-        Queen queen = new Queen(Team.WHITE);
+        Rook rook = new Rook(Team.WHITE);
         List<Coordinate> expected = List.of(
                 new Coordinate(4, 'a'),
                 new Coordinate(4, 'b'),
@@ -38,34 +36,20 @@ class QueenTest {
                 new Coordinate(8, 'd'),
                 new Coordinate(3, 'd'),
                 new Coordinate(2, 'd'),
-                new Coordinate(1, 'd'),
-
-                new Coordinate(1, 'a'),
-                new Coordinate(2, 'b'),
-                new Coordinate(3, 'c'),
-                new Coordinate(3, 'e'),
-                new Coordinate(2, 'f'),
-                new Coordinate(1, 'g'),
-                new Coordinate(5, 'e'),
-                new Coordinate(6, 'f'),
-                new Coordinate(7, 'g'),
-                new Coordinate(8, 'h'),
-                new Coordinate(5, 'c'),
-                new Coordinate(6, 'b'),
-                new Coordinate(7, 'a')
+                new Coordinate(1, 'd')
         );
 
-        List<Coordinate> result = queen.findAllPossibleCoordinate(start);
+        List<Coordinate> result = rook.findAllPossibleCoordinate(start);
 
         assertThat(result)
                 .containsExactlyInAnyOrderElementsOf(expected);
     }
 
-    @DisplayName("퀸이 모서리칸에 존재하는 경우를 검증한다.")
+    @DisplayName("룩이 모서리칸에 존재하는 경우를 검증한다.")
     @Test
     void findAllPossibleCoordinate2() {
         Coordinate start = new Coordinate(1, 'a');
-        Queen queen = new Queen(Team.WHITE);
+        Rook rook = new Rook(Team.WHITE);
         List<Coordinate> expected = List.of(
                 new Coordinate(8, 'a'),
                 new Coordinate(7, 'a'),
@@ -80,21 +64,12 @@ class QueenTest {
                 new Coordinate(1, 'e'),
                 new Coordinate(1, 'f'),
                 new Coordinate(1, 'g'),
-                new Coordinate(1, 'h'),
-
-                new Coordinate(2, 'b'),
-                new Coordinate(3, 'c'),
-                new Coordinate(4, 'd'),
-                new Coordinate(5, 'e'),
-                new Coordinate(6, 'f'),
-                new Coordinate(7, 'g'),
-                new Coordinate(8, 'h')
+                new Coordinate(1, 'h')
         );
 
-        List<Coordinate> result = queen.findAllPossibleCoordinate(start);
+        List<Coordinate> result = rook.findAllPossibleCoordinate(start);
 
         assertThat(result)
                 .containsExactlyInAnyOrderElementsOf(expected);
     }
 }
-
